+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/faq/FetchFAQUseCase.kt
@@ -1,0 +1,18 @@
+package no.usn.mob3000.domain.usecase.content.faq
+
+import no.usn.mob3000.data.repository.content.FAQRepository
+import no.usn.mob3000.domain.model.FAQData
+
+/**
+ * Use case for fetching FAQ.
+ *
+ * @author 258030
+ * @created 2024-10-30
+ */
+class FetchFAQUseCase(
+    private val fetchRepository: FAQRepository = FAQRepository()
+) {
+    suspend fun fetchFAQ(): Result<List<FAQData>> {
+        return fetchRepository.fetchFAQ()
+    }
+}
