+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/news/DeleteNewsUseCase.kt
@@ -1,0 +1,18 @@
+package no.usn.mob3000.domain.usecase.content.news
+
+import no.usn.mob3000.data.repository.content.NewsRepository
+import no.usn.mob3000.domain.repository.INewsRepository
+
+/**
+ * Use case for deleting news.
+ *
+ * @author 258030
+ * @created 2024-10-30
+ */
+class DeleteNewsUseCase(
+    private val newsRepository: INewsRepository = NewsRepository()
+) {
+    suspend fun deleteNews(newsId: String): Result<Unit> {
+        return newsRepository.deleteNews(newsId)
+    }
+}
